Fix generics in custom-set's example solution
Fixes a problem where the to_map output generics argument (int) did not
match the type returned by the mapping function (u8). Also, the generics
arugment to the invert seem to be the wrong way around (according to its
docs).

Suspect this was causing the following error from the compiler in the CI
build:

  cc: /tmp/v_1001/tsession_7f5b1df3c740_01K272WVY8G5PCP208XZECN73S/run_test.01K272WVYC0AT6TAKW0X27673H.tmp.c:1921: error: invalid type

See also:
- https://modules.vlang.io/maps.html#to_map
- https://modules.vlang.io/maps.html#invert